+++ PetShopApp.WebApi/Controllers/PetsController.cs
@@ -23,20 +23,47 @@ namespace PetShopApp.WebApi.Controllers
         [HttpGet]
         public ActionResult<List<Pet>> Get()
         {
-            return _petService.GetAllPets();
+            try
+            {
+                return _petService.GetAllPets();
+            }
+            catch (Exception e)
+            {
+                return StatusCode(500, "Task failed successfully");
+            }
+            
         }
 
         // GET api/<PetsController>/5
         [HttpGet("{id}")]
         public ActionResult<Pet> Get(int id)
         {
-            return _petService.FindPetById(id);
+            var pet = _petService.FindPetById(id);
+            if (pet == null)
+            {
+                return StatusCode(404, "Did not find Pet with ID " + id);
+            }
+
+            try
+            {
+                return _petService.FindPetById(id);
+            }
+            catch (Exception e)
+            {
+                return StatusCode(500, "Task failed successfully");
+            }
+            
         }
 
         // POST api/<PetsController>
         [HttpPost]
         public ActionResult<Pet> Post([FromBody] Pet pet)
         {
+            if (string.IsNullOrEmpty(pet.Name))
+            {
+                return StatusCode(500, "Name is required for Creating Pet");
+            }
+
             return _petService.CreatePet(pet);
         }
 
@@ -44,14 +71,41 @@ namespace PetShopApp.WebApi.Controllers
         [HttpPut("{id}")]
         public ActionResult<Pet> Put(int id, [FromBody] Pet pet)
         {
-            return _petService.UpdatePet(pet);
+            if (id < 1 || id != pet.Id)
+            {
+                return StatusCode(404, "Parameter Id and Pet Id must be the same");
+            }
+
+            try
+            {
+                return _petService.UpdatePet(pet);
+            }
+            catch (Exception e)
+            {
+                return StatusCode(500, "Task failed successfully");
+            }
+            
         }
 
         // DELETE api/<PetsController>/5
         [HttpDelete("{id}")]
         public ActionResult<Pet> Delete(int id)
         {
-            return _petService.DeletePet(id);
+            var pet = _petService.DeletePet(id);
+            if (pet == null)
+            {
+                return StatusCode(404, "Did not find Pet with ID " + id);
+            }
+
+            try
+            {
+                return _petService.DeletePet(id);
+            }
+            catch (Exception e)
+            {
+                return StatusCode(500, "Task failed successfully");
+            }
+            
         }
     }
 }
